server: move Player from model pkg to realtime pkg
